+++ packages/log/package.json
@@ -23,18 +23,18 @@
   },
   "main": "./cjs/index.js",
   "browser": {
-    "./cjs/platform/index.js": "./cjs/platform/index.browser.js",
     "./cjs/getDefaultReporter.js": "./cjs/getDefaultReporter.browser.js",
     "./cjs/getLogModeFromEnvironment.js": "./cjs/getLogModeFromEnvironment.browser.js",
     "./cjs/memory/toInspectStringForObject.ts": "./cjs/memory/toInspectStringForObject.browser.ts",
-    "./esm/platform/index.js": "./esm/platform/index.browser.js",
+    "./cjs/platform/index.js": "./cjs/platform/index.browser.js",
     "./esm/getDefaultReporter.js": "./esm/getDefaultReporter.browser.js",
     "./esm/getLogModeFromEnvironment.js": "./esm/getLogModeFromEnvironment.browser.js",
     "./esm/memory/toInspectStringForObject.ts": "./esm/memory/toInspectStringForObject.browser.ts",
-    "./ts/platform/index.js": "./ts/platform/index.browser.js",
+    "./esm/platform/index.js": "./esm/platform/index.browser.js",
     "./ts/getDefaultReporter.ts": "./ts/getDefaultReporter.browser.ts",
     "./ts/getLogModeFromEnvironment.ts": "./ts/getLogModeFromEnvironment.browser.ts",
-    "./ts/memory/toInspectStringForObject.ts": "./ts/memory/toInspectStringForObject.browser.ts"
+    "./ts/memory/toInspectStringForObject.ts": "./ts/memory/toInspectStringForObject.browser.ts",
+    "./ts/platform/index.js": "./ts/platform/index.browser.js"
   },
   "typings": "./cjs/index.d.ts",
   "files": [
