--- packages/log/ts/console/types.ts
@@ -1,3 +1,0 @@
-import { polyfilledConsole } from './polyfilledConsole.js'
-
-export type Console = typeof polyfilledConsole
