--- packages/log/ts/formatter/formatLogLevel.ts
@@ -1,6 +1,0 @@
-import { upperCase } from 'upper-case'
-import { toLogLevelName } from '../logLevelFn.js'
-
-export function formatLogLevel(level: number) {
-  return `(${upperCase(toLogLevelName(level))})`
-}
