+++ packages/log/ts/standard-log/createStandardLog.ts
@@ -1,0 +1,5 @@
+import { ConfigOptions } from '../config.js'
+
+export function createStandardLog(options: Partial<ConfigOptions>) {
+  return options
+}
